requirements.txt
django-registration
django-countries
python-decouple